Move rest/capnp/* to sm/serialization/ and rest/curl/* into sm/rest/
