+++ test/src/unit-capi-incomplete.cc
@@ -32,9 +32,9 @@
 
 #include "catch.hpp"
 #include "test/src/helpers.h"
-#include "tiledb/rest/capnp/query.h"
 #include "tiledb/sm/c_api/tiledb.h"
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
+#include "tiledb/sm/serialization/query.h"
 
 #include <cstring>
 #include <iostream>
