+++ test/src/unit-capi-rest-dense_array.cc
@@ -42,9 +42,9 @@
 #include "tiledb/sm/filesystem/posix.h"
 #endif
 #include "test/src/helpers.h"
-#include "tiledb/rest/curl/rest_client.h"
 #include "tiledb/sm/c_api/tiledb.h"
 #include "tiledb/sm/misc/utils.h"
+#include "tiledb/sm/rest/rest_client.h"
 
 #include <array>
 #include <cassert>
