+++ tiledb/sm/array/array.cc
@@ -31,11 +31,11 @@
  */
 
 #include "tiledb/sm/array/array.h"
-#include "tiledb/rest/curl/rest_client.h"
 #include "tiledb/sm/encryption/encryption.h"
 #include "tiledb/sm/enums/serialization_type.h"
 #include "tiledb/sm/misc/logger.h"
 #include "tiledb/sm/misc/stats.h"
+#include "tiledb/sm/rest/rest_client.h"
 
 #include <cassert>
 #include <iostream>
