+++ tiledb/sm/serialization/array_schema.cc
@@ -32,7 +32,6 @@
 
 #include "capnp/compat/json.h"
 #include "capnp/serialize.h"
-#include "tiledb/rest/capnp/utils.h"
 #include "tiledb/sm/enums/array_type.h"
 #include "tiledb/sm/enums/compressor.h"
 #include "tiledb/sm/enums/datatype.h"
@@ -41,6 +40,7 @@
 #include "tiledb/sm/misc/constants.h"
 #include "tiledb/sm/misc/logger.h"
 #include "tiledb/sm/misc/stats.h"
+#include "tiledb/sm/serialization/utils.h"
 
 namespace tiledb {
 namespace rest {
